Distributed the entire project throughout different files having related functionality in each
